Code cleanup, lower audio settings
Former-commit-id: 69b6a9728b7850c0213415c92cffadd494167f06

+++ WorldsEngine/Core/Engine.cpp
@@ -171,7 +171,7 @@ namespace worlds {
         if (config) {
             memcpy(config->Name, fontPath.c_str(), fontPath.size());
         } else {
-            for (int i = 0; i < fontPath.size(); i++) {
+            for (size_t i = 0; i < fontPath.size(); i++) {
                 defaultConfig.Name[i] = fontPath[i];
             }
             config = &defaultConfig;
